Prototype for translation via xlsx files.
To export, navigate to the translation menu for an article, copy the URL you
get from the "Download PO File" button, and replace 'pofile' with 'xlsxfile'
and prepend 'translation/' before 'admin/...'.
To import, take the above URL and change 'download' to 'upload'. You will need
to send a POST request to that URL which contains the file you want to upload.